Improve initiate process
